Añadiendo validaciones, busqueda y paginado.

+++ ControlEscolarApi.Application/Personal/Commands/CreatePersonal/CreatePersonalCommand.cs
@@ -11,6 +11,6 @@ public class CreatePersonalCommand : IRequest<ErrorOr<Domain.Entities.Personal>>
     public string Correo { get; set; } = null!;
     public DateTime FechaNacimiento { get; set; }
     public bool Estatus { get; set; }
-    public float Sueldo { get; set; }
+    public decimal Sueldo { get; set; }
     public int TipoPersonalId { get; set; }
 }
