+++ ControlEscolarApi.Application/Personal/Commands/UpdatePersonal/UpdatePersonalCommand.cs
@@ -12,6 +12,6 @@ public class UpdatePersonalCommand : IRequest<ErrorOr<Domain.Entities.Personal>>
     public string? Correo { get; set; } = null!;
     public DateTime? FechaNacimiento { get; set; }
     public bool? Estatus { get; set; }
-    public float? Sueldo { get; set; }
+    public decimal? Sueldo { get; set; }
     public int? TipoPersonalId { get; set; }
 }
