+++ ControlEscolarApi.Application/TiposPersonal/Commands/CreateTipoPersonal/CreateTipoPersonalCommand.cs
@@ -8,6 +8,6 @@ public class CreateTipoPersonalCommand : IRequest<ErrorOr<TipoPersonal>>
 {
   public string Prefijo { get; set;} = null!;
   public string Nombre { get; set;} = null!;
-  public float SueldoMinimo { get; set;}
-  public float SueldoMaximo { get; set;}
+  public decimal SueldoMinimo { get; set;}
+  public decimal SueldoMaximo { get; set;}
 }
