+++ ControlEscolarApi.Application/TiposPersonal/Commands/UpdateTipoPersonal/UpdateTipoPersonalCommand.cs
@@ -8,7 +8,7 @@ public class UpdateTipoPersonalCommand :  IRequest<ErrorOr<TipoPersonal>>{
   public int Id { get; set; }
   public string? Prefijo { get; set;}
   public string? Nombre { get; set;}
-  public float? SueldoMinimo { get; set;}
-  public float? SueldoMaximo { get; set;}
+  public decimal? SueldoMinimo { get; set;}
+  public decimal? SueldoMaximo { get; set;}
 }
