+++ ControlEscolarApi.Contracts/Personal/CreatePersonalRequest.cs
@@ -8,6 +8,6 @@ public class CreatePersonalRequest
     public string Correo { get; set; } = null!;
     public DateTime FechaNacimiento { get; set; }
     public bool Estatus { get; set; }
-    public float Sueldo { get; set; }
+    public decimal Sueldo { get; set; }
     public int TipoPersonalId { get; set; }
 }
