+++ ControlEscolarApi.Contracts/TiposPersonal/TipoPersonalResponse.cs
@@ -4,6 +4,6 @@ public class TipoPersonalResponse {
   public int Id { get; set; }
   public string Prefijo { get; set;} = null!;
   public string Nombre { get; set;} = null!;
-  public float SueldoMinimo { get; set;}
-  public float SueldoMaximo { get; set;}
+  public decimal SueldoMinimo { get; set;}
+  public decimal SueldoMaximo { get; set;}
 }
