+++ ControlEscolarApi.Contracts/TiposPersonal/UpdateTipoPersonalRequest.cs
@@ -5,6 +5,6 @@ public class UpdateTipoPersonalRequest {
   public int Id { get; set; }
   public string? Prefijo { get; set;}
   public string? Nombre { get; set;}
-  public float? SueldoMinimo { get; set;}
-  public float? SueldoMaximo { get; set;}
+  public decimal? SueldoMinimo { get; set;}
+  public decimal? SueldoMaximo { get; set;}
 }
