+++ ControlEscolarApi.Domain/Entities/TipoPersonal.cs
@@ -5,6 +5,6 @@ public class TipoPersonal
     public int Id { get; set;}
     public string Prefijo { get; set; } = null!;
     public string Nombre { get; set;} = null!;
-    public float SueldoMinimo { get; set;}
-    public float SueldoMaximo { get; set;}
+    public decimal SueldoMinimo { get; set;}
+    public decimal SueldoMaximo { get; set;}
 }
